run-test.sh outputs details of a failed test
informing the user of the test run results needs to be
a split responsibility to plug the SPoF in the exit code
of run-test.sh.  if either run-test.sh or run-tests.sh fails
to convey results correctly, the overall result would be
obviously inconsistent, which is better than *only apparently*
passing tests.